Solved some x64 errors and warnings and some general warnings

+++ zbar/config.c
@@ -36,14 +36,14 @@ int zbar_parse_config (const char *cfgstr,
                        int *val)
 {
     const char *dot, *eq;
-    int len;
+    ptrdiff_t len;
     char negate;
     if(!cfgstr)
         return(1);
 
     dot = strchr(cfgstr, '.');
     if(dot) {
-        int len = dot - cfgstr;
+        ptrdiff_t len = dot - cfgstr;
         if(!len || (len == 1 && !strncmp(cfgstr, "*", len)))
             *sym = 0;
         else if(len < 2)
